% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{A_to_array}
\alias{A_to_array}
\title{Transform matrix A(p x p*d) to array p times p times d}
\usage{
A_to_array(A)
}
\arguments{
\item{A}{matrix of size p times p*d}
}
\value{
array of size p times p times d
}
\description{
Transform matrix A (p times p*d) to array p times p times d
}
