% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{adaptive_bandwidth_TLM}
\alias{adaptive_bandwidth_TLM}
\title{Adaptive rule to chose the lag-window}
\usage{
adaptive_bandwidth_TLM(
  X,
  Kn = 5,
  c.thres = 2,
  acf.max = (dim(X)[1])^(1/2),
  even = T,
  Kern = Kernel
)
}
\arguments{
\item{X}{Multivariate time series}

\item{Kn}{Tuning parameter Kn discussed in Politis (2003). Roughly, the number of lags the autocorrelation function must stay below the threshold determined by c.thres}

\item{c.thres}{The bandwidth is chosen by looking for the first time the Frobenius norm of the autocorrelation function drops below \code{c.thres * p*sqrt(1/n)} and stays below that level for \code{Kn} lags.}

\item{acf.max}{Maximum number of autocorrelations to consider}

\item{even}{flag if always and even lag window should be returned}

\item{Kern}{used kernel}
}
\value{
list: M: suggested lag-window, M_corr_fac: Scaling factor for used kernel considered Jenkins kernel-bandwidth adjustment
}
\description{
Adaptive rule to chose the lag-window, see Politis (2003): Adaptive bandwidth choice. This code is a modification of iosmooth::bwadap.ts by T. McMurry.
here, the Frobenius-norm of the autocorrelation matrix are considered instead of the autocorrelation itself. The threshold is scaled by a factor p.
}
