% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{Array_to_A}
\alias{Array_to_A}
\title{Transform array p times p times d  to matrix A(p times p*d)}
\usage{
Array_to_A(A)
}
\arguments{
\item{A}{array of size p times p times d}
}
\value{
matrix of size p times p*d
}
\description{
Transform array p times p times d  to matrix A(p times p*d)
}
