% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.r
\name{Emp_power}
\alias{Emp_power}
\title{Computes the empirical power}
\usage{
Emp_power(TX, t_hat, True_R, Delta_0)
}
\arguments{
\item{TX}{Test statistic}

\item{t_hat}{Chosen threshold}

\item{True_R}{True (partial) coherence}

\item{Delta_0}{Delta in H0, i.e., (partial) coherence above that delta shall be identified}
}
\value{
empirical power
}
\description{
Computes the empirical power for detecting (partial) coherences above the threshold Delta_0
}
