% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{f_inv_to_part_rho}
\alias{f_inv_to_part_rho}
\title{f_inv_to_part_rho}
\usage{
f_inv_to_part_rho(f_0)
}
\arguments{
\item{f_0}{\eqn{f^{-1}} as array of size p times p times d, where p is the dimension and d is the number of frequencies}
}
\value{
\eqn{\rho}  array of size p times p times d, where p is the dimension and d is the number of frequencies
}
\description{
computes \eqn{\rho} of the partial coherence given \eqn{f^{-1}} for all frequencies of \eqn{f^{-1}}.
}
