% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{f_spec_inv_ARMA}
\alias{f_spec_inv_ARMA}
\title{Spectral density matrix inverse at frequency Omega for a VARMA(d,q) model}
\usage{
f_spec_inv_ARMA(A, B, Sigma, omega_list)
}
\arguments{
\item{A}{VAR model matrix as matrix p times p*VAR order}

\item{B}{VMA model matrix as matrix p times p*VMA order}

\item{Sigma}{Sigma  matrix as p times p matrix}

\item{omega_list}{vector of frequencies}
}
\value{
spectral density inverse matrix as array of size p times p times length(omega_list)
}
\description{
Compute the inverse spectral density at frequences Omega for a VARMA(d,q) model with parameter A (VAR), B (VMA) and variance matrix of the residuals Sigma.
}
