% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{f_spec_MA}
\alias{f_spec_MA}
\title{Spectral density matrix inverse at frequency Omega for a VMA(d) model}
\usage{
f_spec_MA(A, Sigma, omega_list)
}
\arguments{
\item{A}{VMA model matrix as matrix p times p*MA order}

\item{Sigma}{Sigma  matrix as p times p matrix}

\item{omega}{vector of frequencies}
}
\value{
spectral density inverse matrix as array of size p times p times length(Omega)
}
\description{
Compute the inverse spectral density at frequences Omega for a VMA(d) model with parameter A and variance matrix of the residuals Sigma.
}
