% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{f_to_f_inv}
\alias{f_to_f_inv}
\title{f_to_f_inv}
\usage{
f_to_f_inv(f_0)
}
\arguments{
\item{f_0}{\eqn{f^{-1}} as array of size p times p times d, where p is the dimension and d is the number of frequencies}
}
\value{
\eqn{f^{-1}} array of size p times p times d, where p is the dimension and d is the number of frequencies
}
\description{
computes \eqn{f^{-1}} given f for all frequencies of \eqn{f}
}
