% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{f_to_Gamma}
\alias{f_to_Gamma}
\title{f_to_Gamma}
\usage{
f_to_Gamma(f_0, h)
}
\arguments{
\item{f_0}{\eqn{f} as array of size p times p times d, where p is the dimension and d is the number of frequencies}

\item{h}{lag for the autocovariance}
}
\value{
\eqn{\Gamma} as array of size p times p
}
\description{
Computes autocovariance \eqn{\Gamma} to lag h given spectral denstiy f_0
}
