% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{f_to_norm_f}
\alias{f_to_norm_f}
\title{Coherence}
\usage{
f_to_norm_f(f)
}
\arguments{
\item{f}{\eqn{f} as array of size p times p times d, where p is the dimension and d is the number of frequencies}
}
\value{
coherence as array of size p times p times d, where p is the dimension and d is the number of frequencies
}
\description{
computes the coherence given \eqn{f} for all frequencies of \eqn{f}.
}
