% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.r
\name{FDR_TX}
\alias{FDR_TX}
\title{Computes the empirical false discovery rate}
\usage{
FDR_TX(TX, t_hat, True_R, Delta_0)
}
\arguments{
\item{TX}{Test statistic}

\item{t_hat}{Chosen threshold}

\item{True_R}{True (partial) coherence}

\item{Delta_0}{Delta in H0, i.e., (partial) coherence above that delta shall be identified}
}
\value{
empirical FDR
}
\description{
Computes the empirical false discovery rate for detecting (partial) coherences above the threshold Delta_0
}
