% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{Gamma_to_f}
\alias{Gamma_to_f}
\title{Gamma_to_f}
\usage{
Gamma_to_f(Gamma_dn_array, omega)
}
\arguments{
\item{Gamma_dn_array}{array of size p times p times d, where p is the dimension and d is the number of autocovariance. The lags need to be stored as lag=0,1,...,d-1}

\item{omega}{frequency, where the spectral density matrix should be computed}
}
\value{
spectral density matrix at frequency omega as p times p matrix.
}
\description{
Computes the spectral density matrix at frequency omega given an array of autocovariance matrices (lag=0,1,...,d-1).
}
