% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularized_regression.R
\name{glasso_BIC}
\alias{glasso_BIC}
\title{Graphical lasso with BIC}
\usage{
glasso_BIC(S, n, lam, maxit = 10000, tol = 1e-04, gamma = 4)
}
\arguments{
\item{S}{Input matrix, i.e., empirical covariance matrix or (transformed) lag-window spectral density matrix}

\item{n}{(Effective) sample size (important for BIC)}

\item{lam}{regularization parameter lambda}

\item{maxit}{maximal number of iterations}

\item{tol}{Tolerance in numerical optimization}

\item{gamma}{Gamma-Parameter in BIC selection with considerations of Wang. et. al (2009)}
}
\value{
list: Lambdas: used lambda seqeuence, maxIt: maximal number of iterations,Omega: Estimated precision matrix or f^{-1},Sigma: Estimated Covariance matrix of f,BICs: BIC values, best_lam: BIC-optimal lambda
}
\description{
This is a wrapper for the glassoFast function to compute the graphical lasso.
Additionally, the tuning parameter of the regularization parameter lambda is chosen by BIC (taking the following considerations into account: Wang, H., Li, B., and Leng, C. (2009). Shrinkage tuning parameter selection with a diverging number of parameters)
The (real-valued) input matrix S is standardized prior to computation.
}
