% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularized_regression.R
\name{glasso_seq_lam}
\alias{glasso_seq_lam}
\title{Graphical lasso}
\usage{
glasso_seq_lam(S, n, lam, maxit = 10000, tol = 1e-04)
}
\arguments{
\item{S}{Input matrix, i.e., empirical covariance matrix or (transformed) lag-window spectral density matrix}

\item{n}{(Effective) sample size (important for BIC)}

\item{lam}{regularization parameter lambda}

\item{maxit}{maximal number of iterations}

\item{tol}{Tolerance in numerical optimization}
}
\value{
list: Lambdas: used lambda seqeuence, Omega_seq: Sequences (length(lam)) of estimated precision matrix or f^{-1}
}
\description{
This is a wrapper for the glassoFast function to compute the graphical lasso over a sequence of regularization parameters lambda.
The (real-valued) input matrix S is standardized prior to computation.
}
