% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.r
\name{Gt2t}
\alias{Gt2t}
\title{The Cumulative distribution function of d independent chi2 distributions}
\usage{
Gt2t(t, d)
}
\arguments{
\item{t}{vector of value at which probabilities shall be computed}

\item{d}{Number of independent chi2 random variables}
}
\value{
probability at t
}
\description{
The Cumulative distribution function of d independent \eqn{chi^2}-distributed random variables
}
