% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{max_coh_H_0_pre_Filt}
\alias{max_coh_H_0_pre_Filt}
\title{Test statistic based on the maximal coherence with pre-filtering}
\usage{
max_coh_H_0_pre_Filt(
  Z,
  Kernel,
  M,
  DEMEAN = TRUE,
  Delta_0,
  FourierFrequencies_h,
  Filter
)
}
\arguments{
\item{Z}{Discrete Fourier transformed as matrix n (sample size) times p (dimension)}

\item{Kernel}{used kernel in frequency domain as a function with input frequencies and lag window}

\item{M}{lag window for kernel}

\item{DEMEAN}{flag if Z should be centered prior to kernel estimation}

\item{Delta_0}{specify Delta 0 is test is not performed to Delta_0=0.}

\item{FourierFrequencies_h}{Fourier Frequencies at which the de-biased partial coherence should be computed}

\item{Filter}{Filter for the pre-filtering step. Filter input as a VAR filter as matrix of size p times (p*VARorder)}
}
\value{
list: TX_Z: test statistic in raw form; TX_chi: test statistic in chi^2 form; coh_rho: partial coherence; C_K: kernel variance constant; M: used bandwidth
}
\description{
Computes the test statistic based on the maximum coherences over the frequencies (FourierFrequencies_h).
The coherence is computed with the pre-filtered lag-window coherence estimator.
}
