% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{max_de_part_coh_H_0}
\alias{max_de_part_coh_H_0}
\title{Test statistic based on the maximal de-biased partial coherence}
\usage{
max_de_part_coh_H_0(
  Z,
  spec,
  Kernel,
  M,
  DEMEAN = TRUE,
  Delta_0,
  FourierFrequencies_h
)
}
\arguments{
\item{Z}{Discrete Fourier transformed as matrix n (sample size) times p (dimension)}

\item{spec}{estimated f^{-1} at frequencies Frequencies of size p times p times length(Frequencies)}

\item{Kernel}{used kernel in frequency domain as a function with input frequencies and lag window}

\item{M}{lag window for kernel}

\item{DEMEAN}{flag if Z should be centered prior to kernel estimation}

\item{Delta_0}{specify Delta 0 is test is not performed to Delta_0=0.}

\item{FourierFrequencies_h}{Fourier Frequencies at which the de-biased partial coherence should be computed}
}
\value{
list: TX_Z: test statistic in raw form; TX_chi: test statistic in chi^2 form; part_rho: de-biased partial coherence; C_K: kernel variance constant; M: used bandwidth
}
\description{
Computes the test statistic based on the maximum partial coherences over the frequencies (FourierFrequencies_h).
The partial coherence is computed with the de-biased partial coherence estimator.
}
