% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{Modified_Bartlett}
\alias{Modified_Bartlett}
\title{Modified Bartlett kernel}
\usage{
Modified_Bartlett(lambda, M)
}
\arguments{
\item{lambda}{input frequencies}

\item{M}{window length}
}
\value{
kernel weights
}
\description{
Modified Bartlett kernel (triangular kernel \eqn{K(u)=1-|u|} for \eqn{u \in (-1,1)}) with lag window M written in frequency domain.
Note, this kernel is positive definite.
}
