% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{part_coh_de_Filter}
\alias{part_coh_de_Filter}
\title{De-biased partial coherence with pre-filtering
Computes for frequencies Frequencies the de-biased partial coherence (rho) for all u,v.
A VAR-prefiltering (Filter) is used in the kernel estimation.
spec should be an estimate of f^{-1} at frequencies Frequencies.}
\usage{
part_coh_de_Filter(Z, spec, Filter, Frequencies, Kernel, M, DEMEAN = FALSE)
}
\arguments{
\item{Z}{Discrete Fourier transformed as matrix n (sample size) times p (dimension)}

\item{spec}{estimated f^{-1} at frequencies Frequencies of size p times p times length(Frequencies)}

\item{Filter}{Filter for the pre-filtering step. Filter input as a VAR filter as matrix of size p times (p*VARorder)}

\item{Frequencies}{Frequencies at which the de-biased partial coherence should be computed}

\item{Kernel}{used kernel in frequency domain as a function with input frequencies and lag window}

\item{M}{lag window for kernel}

\item{DEMEAN}{flag if Z should be centered prior to kernel estimation}
}
\value{
rho_{u,v}(omega), omega in Frequencies, u,v=1,...,p  as array of p times p times length(Frequencies)
}
\description{
De-biased partial coherence with pre-filtering
Computes for frequencies Frequencies the de-biased partial coherence (rho) for all u,v.
A VAR-prefiltering (Filter) is used in the kernel estimation.
spec should be an estimate of f^{-1} at frequencies Frequencies.
}
