% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{PC_Network}
\alias{PC_Network}
\title{Computes the (partial) coherence network}
\usage{
PC_Network(
  X_t,
  Frequencyband = c(0, pi),
  FDR_alpha = 0.05,
  Delta_0 = 0,
  Kernel = Modified_Bartlett,
  r = 2,
  lambda = NULL,
  VAR_order = ceiling(log(dim(X_t)[1], 10)),
  bandwidth = NULL,
  mbic2_c = log(dim(X_t)[2]),
  gamma = log(log(dim(X_t)[2]))
)
}
\arguments{
\item{X_t}{time series}

\item{Frequencyband}{lower and upper value of the frequency band (subset of (0,pi)) to consider}

\item{FDR_alpha}{desired FDR}

\item{Delta_0}{Delta in H0, i.e., (partial) coherence above that delta shall be identified. Delta0=0 means all non-zero (partial) coherences}

\item{Kernel}{Used kernel}

\item{r}{r is determined by the decay behavior of the Fourier coefficients of the kernel used; see Assumption 2 in Section 6. r=NULL for uniform kernel, r=2 for modified Bartlett}

\item{lambda}{lambda sequence for regularization}
}
\value{
list: Frequencyband_Fourier: Used Fourier frequencies; Partial_Coherence: Partial coherence network, Coherence: Coherence network
}
\description{
This function computes for a given time series X_t its coherence and partial coherence network. The objective is to identify all (partial) coherences which have values above Delta_0 over the frequency band (Frequencyband).
For this, first the spectral density matrix inverse is estimated using a pre-whitten (sparse VAR) lag-window estimator (with kernel Kernel) with graphical lasso regularization. Then, the de-biased partial coherence (with kernel Kernel) is computed and the test statistics
are threshold automatically to obtain (asymptotically) the FDR FDR_alpha.
}
