% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularized_regression.R
\name{Regression_Row_IC}
\alias{Regression_Row_IC}
\title{Given Regression Y=MODEL(i) X , computes several information criteria  and selects the best MODEL i}
\usage{
Regression_Row_IC(
  Y,
  X,
  MODEL,
  crit = c("bic", "aic", "aicc", "hqc", "mbic2", "eric", "bic_m"),
  lambda,
  mbic2_c = 4
)
}
\arguments{
\item{Y}{vector y of length n}

\item{X}{matrix X of size n times p}

\item{MODEL}{model parameter Beta ( Y=BETA X ) as array p times length(lambda)}

\item{crit}{flag to specify which information criteria should be used}

\item{lambda}{vector of tuning paramter lambda}

\item{mbic2_c}{additinal parameter for some information criteria}
}
\value{
list: coef: Best model according to used IC; eps: Residuals of best model; lam: optimal lambda accoirding to used IC; crit: Minimal value of used IC.
}
\description{
Given Regression Y=MODEL(i) X , computes several information criteria  and selects the best MODEL i.
Information criteria: (n sample size, p dimension, n_var number of non-zero variables)
bic = n * log(mse) + nvar * log(n)
mbic2 = n * log(mse) + nvar * (log(n)+log(p/mbic2_c))-2\emph{log(gamma(nvar+1))
aic = n * log(mse) + 2 * nvar
aicc = aic + (2 * nvar * (nvar + 1))/(n - nvar - 1)
hqc = n * log(mse) + 2 * nvar * log(log(n))
eric=n * log(mse) +nvar * log(sqrt(n}mse)/lambda)*ifelse(n>p,1,0.5)
bic_m = n * log(mse) + nvar * log(n)*mbic2_c
}
