% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{S_comp}
\alias{S_comp}
\title{S_comp
Given discrete Fourier transformed Z, computes \eqn{Z Z^*} and splits it into real and imaginary parts.}
\usage{
S_comp(Zt)
}
\arguments{
\item{Zt}{complex matrix of size n times p}
}
\value{
real matrix of size 2p times 2p
}
\description{
S_comp
Given discrete Fourier transformed Z, computes \eqn{Z Z^*} and splits it into real and imaginary parts.
}
