% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{simulateVARMAs}
\alias{simulateVARMAs}
\title{Simulates a VARMA model}
\usage{
simulateVARMAs(A, B, Sigma, n, n2 = 100)
}
\arguments{
\item{A}{VAR model matrix as matrix p times p*VAR order}

\item{B}{VMA model matrix as matrix p times p*VMA order}

\item{Sigma}{Sigma  matrix as p times p matrix}

\item{n}{number of samples to generate}

\item{n2}{burn-in period}
}
\value{
multivariate time series X_t as matrix n times p
}
\description{
Generates n realization of a p-dimensional VARMA(VAR order,VMA order) model.
}
