% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{spec_inv_freq}
\alias{spec_inv_freq}
\title{Spectral density matrix inverse estimator}
\usage{
spec_inv_freq(
  Z,
  lambda,
  Kernel,
  M,
  Frequencies,
  THR = FALSE,
  Check_pos_def = T,
  M_corr = 1,
  gamma = 0.5,
  ...
)
}
\arguments{
\item{Z}{Discrete Fourier transformed as matrix n (sample size) times p (dimension)}

\item{lambda}{tuning parameter for graphical lasso}

\item{Kernel}{used kernel in frequency domain as a function with input frequencies and lag window}

\item{M}{lag window for kernel}

\item{Frequencies}{frequencies at which f^{-1} should be computed}

\item{THR}{flag if output should be thresholded at lambda}

\item{Check_pos_def}{flag if output is transformed to positive definiteness}

\item{M_corr}{Jenkins kernel-bandwidth adjustmentl. I.e., scaling to same variance as uniform kernel}

\item{gamma}{tuning parameter for ebic. Gamma=0 results in classical BIC}

\item{...}{optional parameters for graphical lasso.}
}
\value{
list with array of f^{-1} of size p times p times length(Frequencies) and optimal (eBIC) lambdas.
}
\description{
Discrete Fourier transformed Z are used to compute the spectral density matrix inverse f^{-1} at frequency Frequencies.
A graphical lasso (glassofast) is used and tuning parameter lambda is chosen by (e)BIC
Kernel smoothing is used with kernel Kernel and bandwith M
}
