% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{spec_inv_freq_lam}
\alias{spec_inv_freq_lam}
\title{Spectral density matrix inverse estimator}
\usage{
spec_inv_freq_lam(Z, lambda, Kernel, M, Frequencies)
}
\arguments{
\item{Z}{Discrete Fourier transformed as matrix n (sample size) times p (dimension)}

\item{lambda}{tuning parameter for graphical lasso}

\item{Kernel}{used kernel in frequency domain as a function with input frequencies and lag window}

\item{M}{lag window for kernel}

\item{Frequencies}{frequencies at which f^{-1} should be computed}
}
\value{
array of f^{-1} of size p times p times length(Frequencies) times length (lambda)
}
\description{
Discrete Fourier transformed Z are used to compute the spectral density matrix inverse f^{-1} at frequency Frequencies.
A graphical lasso (glassofast) is computed for every tuning parameter lambda.
Kernel smoothing is used with kernel Kernel and bandwith M
}
