% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{spec_inv_preWhiteningVAR_freq}
\alias{spec_inv_preWhiteningVAR_freq}
\title{Spectral density matrix inverse estimator with prewhiteing}
\usage{
spec_inv_preWhiteningVAR_freq(
  X_t,
  lambda,
  Kernel,
  M,
  M_corr = 1,
  VAR_order,
  Frequencies,
  crit = "bic",
  Check_pos_def = T,
  lambda_VAR = lambda,
  gamma = 0.5,
  ...
)
}
\arguments{
\item{X_t}{multivariate time sereis, matrix n times p}

\item{lambda}{tuning parameter for graphical lasso}

\item{Kernel}{used kernel in frequency domain as a function with input frequencies and lag window}

\item{M}{lag window for kernel}

\item{M_corr}{Jenkins kernel-bandwidth adjustmentl. I.e., scaling to same variance as uniform kernel}

\item{VAR_order}{order of the used VAR model}

\item{Frequencies}{frequencies at which f^{-1} should be computed}

\item{crit}{used information criteria to select the tuning parameter lambda in the sparse VAR estimation}

\item{Check_pos_def}{flag if output is transformed to positive definiteness}

\item{lambda_VAR}{tuning parameter lambda in the sparse VAR estimation}

\item{gamma}{tuning parameter for ebic. Gamma=0 results in classical BIC}

\item{...}{optional parameter pased to the sparse VAR estimation}
}
\value{
list with f_inv: array of f^{-1} of size p times p times length(Frequencies); VAR_A: VAR model; f_eps_inv f^{-1} of the VAR residuals; M: bandwidth of f_eps_inv.
}
\description{
An estimator for the spectral density inverse matrix. For this, first a sparse VAR is fitted to the data. Then on the residuals of the sparse VAR,
a nonparameteric spectral density estimator using graphical lasso with eBIC selection is fitted to compute the spectral density matrix inverse of the residuals.
Kernel smoothing is used with kernel Kernel and bandwith M.
Final output is the computed spectral density matrix inverse estimator. I.e, an estimator of \eqn{f^{-1}} of the multivariate time series X_t.
}
