% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral.R
\name{Spec_inv_VAR}
\alias{Spec_inv_VAR}
\title{Spectral density matrix inverse at frequency Omega for a VAR(d) model}
\usage{
Spec_inv_VAR(omega, A, Sigma_inv = NULL, Sigma = NULL)
}
\arguments{
\item{omega}{vector of frequencies}

\item{A}{VAR model matrix as matrix p times p*VARorder}

\item{Sigma_inv}{Sigma inverse matrix as p times p matrix}

\item{Sigma}{Sigma  matrix as p times p matrix}
}
\value{
spectral density inverse matrix as array of size p times p times length(Omega)
}
\description{
Compute the inverse spectral density at frequences Omega for a VAR(d) model with parameter A and variance matrix of the residuals Sigma.
Either Sigma or Sigma inverse is required.
}
