% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.r
\name{t_inf}
\alias{t_inf}
\title{Computes threshold to FDR alpha}
\usage{
t_inf(TX, Gt, q, ttilde, lengt.out, alpha)
}
\arguments{
\item{TX}{Test statistic}

\item{Gt}{Asymptotic distribution of test statistic}

\item{q}{Number of test}

\item{ttilde}{Upper threshold to which point the asymptotic distribution of test statistic is trustworthy}

\item{lengt.out}{Number of threshold to try in interval (0,ttilde)}

\item{alpha}{desired FDR}
}
\value{
Derived threshold
}
\description{
Computes threshold to FDR alpha
}
