% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.r
\name{t_inf_help}
\alias{t_inf_help}
\title{Helper function to determine threshold for test statistic}
\usage{
t_inf_help(t, TX, Gt, q)
}
\arguments{
\item{t}{possible threshold t}

\item{TX}{Test statistic}

\item{Gt}{Asymptotic distribution of test statistic}

\item{q}{Number of test}
}
\value{
Given q conducted tests, it gives the ratio between the expected number of positive test (under H0, i.e., distribution Gt) at threshold t
and the number positive test at threshold t
}
\description{
This is a helper function to determine the threshold for the test statistic (TX) with asymptotic distribution (Gt).
Given q conducted tests, it gives the ratio between the expected number of positive test (under H0, i.e., distribution Gt) at threshold t
and the number positive test at threshold t
}
