% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{Uniform}
\alias{Uniform}
\title{Uniform kernel}
\usage{
Uniform(lambda, M)
}
\arguments{
\item{lambda}{input frequencies}

\item{M}{window length}
}
\value{
kernel weights
}
\description{
Uniform kernel with lag window M written in frequency domain.
Note, this kernel is not positive definite.
}
