% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{VAR_Residuals}
\alias{VAR_Residuals}
\title{Computes the Residuals of a VAR Model
Computes the Residuals of a VAR Model}
\usage{
VAR_Residuals(X_t, A)
}
\arguments{
\item{X_t}{Time Series X_t (n x p)}

\item{A}{VAR model parameter matrix A  as matrix p x (VAR_order * p)}
}
\value{
Residuals as matrix (n-VAR_order)times p
}
\description{
Computes the Residuals of a VAR Model
Computes the Residuals of a VAR Model
}
