% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{VARLASSO}
\alias{VARLASSO}
\title{Estimates a sparse VAR model}
\usage{
VARLASSO(
  X_t,
  VAR_order,
  Variablen = NULL,
  lambda,
  THR = FALSE,
  SCALE = TRUE,
  ADAPTIVE = TRUE,
  tau = 1,
  mbic2_c = 4,
  crit = c("bic", "aic", "aicc", "hqc", "mbic2", "eric", "bic_m"),
  first.penalty.factor = NULL,
  ...
)
}
\arguments{
\item{X_t}{multivariate times series as matrix n times p}

\item{VAR_order}{Order of the VAR model}

\item{Variablen}{vector of indices for which the row-wise VAR model should be computed if interest is only in a subset.}

\item{lambda}{vector of tuning paramter lambda}

\item{THR}{flag to specify if thresholded (with lambda as threshold) VAR model should be returned.}

\item{SCALE}{flag to specify if X should be normalized prior to regularized regression}

\item{ADAPTIVE}{flag to specify if a second adaptive step should be performed}

\item{tau}{if a second adptive is used, tau specifies the following penalty.factor=1/abs(abs(first.step.coef)+1/sqrt(n))^tau}

\item{mbic2_c}{additinal parameter for some information criteria}

\item{crit}{flag to specify which information criteria should be used}

\item{first.penalty.factor}{penalty factors for the first step. NULL leads to uninformative first step factors.}

\item{...}{additional parameters passed to glmnet}
}
\value{
list: A: sparse VAR model as matrix p times p*VARorder; eps: VAR residuals as matrix (n-VAR_order)times p; lambda: selected lambda by IC as vector of length p.
}
\description{
Estimates a sparse VAR model using a row-wise lasso (glmnet). The tuning parameter in the lasso estimation can be chosen by various information criteria, see also see also \code{\link{Regression_row}}
}
