% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{X_to_Z}
\alias{X_to_Z}
\title{X to Z}
\usage{
X_to_Z(X_t)
}
\arguments{
\item{X_t}{n times p dimensional matrix. Multivariate time series.}
}
\value{
Z n times p dimensional complex valued matrix. Finite fourer transform.
}
\description{
Finite discrete Fourier transform of multivariate time series \eqn{X_t=(X_1,...,X_n)}.
Finite discrete Fourier transform \eqn{Z} is to scaling factor \eqn{1/\sqrt(2 \pi n)}.
}
\details{
n sample size
p dimension
}
