% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{Z_to_X}
\alias{Z_to_X}
\title{Z to X}
\usage{
Z_to_X(Z)
}
\arguments{
\item{Z}{n times p dimensional complex valued matrix. Finite fourer transform.}
}
\value{
Z n times p dimensional complex valued matrix. Multivariate time series.
}
\description{
Transforms finite discrete Fourier transform \eqn{Z} of multivariate time series \eqn{X_t=(X_1,...,X_n)} back to its multivariate time series
}
\details{
n sample size
p dimension
}
